ignore unsupported git options
It seems that git can send all sorts of options to the remote
backends.  We don't support any, yet we should at least not crash.

The example I stumbled upon is `push.sign=ifAsked` in the config,
which translated into the `--sign=ifAsked` option, as in [this bug
report](https://git-annex.branchable.com/bugs/gcrypt_repository_not_found/). I
haven't tested if it works and the code is rather ugly, but let's take
this as a proof of concept for now.