Add git-config option to set rsync put flags
Provide the ability to pass flags to `rsync` when uploading.

There are two ways to set the configuration:

- `gcrypt.rsync-put-flags`
- `remote.<name>.gcrypt-rsync-put-flags`

The flags will be applied to `rsync` during uploading when using the `rsync://`
backend. If `remote.<name>.gcrypt-rsync-put-flags` is set, the flags set in
`gcrypt.rsync-put-flags` will not be applied to the remote `<name>`.

This change also includes documentation.

We have tested with the following configurations:

1. none set
2. `git config gcrypt.rsync-put-flags "--perms --chmod=g+rX"`
3. `git config remote.<name>.rsync-put-flags "--perms --chmod=o+rX"`
4. both (2) and (3)

All local files start with only owner permissions set, and umask is set to 077.
In (1), no change in behavior as before, as expected. In (2), the remote files
have the group permissions set, as expected. In (3), the remote files have the
other permissions set, as expected. In (4), the remote files have the other
permissions set, but do not have the group permissions set, as expected.

Signed-off-by: Travis Chen <travis.chen@everchanging.dev>